Moving non index functions from index.h to data.h
