+++ include/af/data.h
@@ -172,8 +172,157 @@ namespace af
     AFAPI array diag(const array &in, const int num = 0, const bool extract = true);
 
     /**
-      @}
+       @}
     */
+
+    /**
+       \defgroup manip_func_join join
+       @{
+       Join two arrays along specified dimension
+
+       \param[in] dim is the dimension along which join occurs
+       \param[in] first is the first input array
+       \param[in] second is the first second array
+       \return the array that joins \p first and \p second along \p dim
+
+       \ingroup manip_mat
+       \ingroup arrayfire_func
+    */
+    AFAPI array join(const int dim, const array &first, const array &second);
+
+    AFAPI array join(const int dim, const array &first, const array &second, const array &third);
+
+    AFAPI array join(const int dim, const array &first, const array &second,
+                     const array &third, const array &fourth);
+    /**
+       @}
+    */
+
+   /**
+      \defgroup manip_func_tile tile
+      @{
+      Tile the input array along specified dimensions
+
+      \param[in] in is the input matrix
+      \param[in] x is the number of times \p in is tiled along first dimension
+      \param[in] y is the number of times \p in is tiled along second dimension
+      \param[in] z is the number of times \p in is tiled along third dimension
+      \param[in] w is the number of times \p in is tiled along fourth dimension
+      \return the tiled output
+
+      \ingroup manip_mat
+      \ingroup arrayfire_func
+   */
+    AFAPI array tile(const array &in, const unsigned x, const unsigned y=1,
+                     const unsigned z=1, const unsigned w=1);
+    /**
+       @}
+    */
+
+    /**
+       \defgroup manip_func_tile tile
+       @{
+       Tile the input array along specified dimensions
+
+       \param[in] in is the input matrix
+       \param[in] dims dim4 of tile dimensions
+       \return the tiled output
+
+       \ingroup manip_mat
+       \ingroup arrayfire_func
+    */
+    AFAPI array tile(const array &in, const dim4 &dims);
+    /**
+       @}
+    */
+
+    /**
+       \defgroup manip_func_reorder reorder
+       @{
+       Reorder the input by in the specified order
+
+       \param[in] in is the input matrix
+       \param[in] x specifies which dimension should be first
+       \param[in] y specifies which dimension should be second
+       \param[in] z specifies which dimension should be third
+       \param[in] w specifies which dimension should be fourth
+       \return the reordered output
+
+       \ingroup manip_mat
+       \ingroup arrayfire_func
+    */
+    AFAPI array reorder(const array& in, const unsigned x,
+                        const unsigned y=1, const unsigned z=2, const unsigned w=3);
+    /**
+       @}
+    */
+
+    /**
+       \defgroup manip_func_shift shift
+       @{
+       Circular shift slong specified dimensions
+
+       \param[in] in is the input matrix
+       \param[in] x specifies the shift along first dimension
+       \param[in] y specifies the shift along second dimension
+       \param[in] z specifies the shift along third dimension
+       \param[in] w specifies the shift along fourth dimension
+       \return the shifted output
+
+       \ingroup manip_mat
+       \ingroup arrayfire_func
+    */
+    AFAPI array shift(const array& in, const int x, const int y=0, const int z=0, const int w=0);
+    /**
+       @}
+    */
+
+    /**
+       \defgroup manip_func_moddims moddims
+       @{
+
+       Modify the input dimensions without changing the data order
+
+       \ingroup manip_mat
+       \ingroup arrayfire_func
+    */
+    AFAPI array moddims(const array& in, const unsigned ndims, const dim_type * const dims);
+
+    AFAPI array moddims(const array& in, const dim4& dims);
+
+    AFAPI array moddims(const array& in, dim_type d0, dim_type d1=1, dim_type d2=1, dim_type d3=1);
+    /**
+       @}
+    */
+
+   /**
+      \defgroup manip_func_flat flat
+      @{
+
+      Flatten the input to a single dimension
+
+      \ingroup manip_mat
+      \ingroup arrayfire_func
+   */
+    AFAPI array flat(const array &in);
+    /**
+       @}
+    */
+
+   /**
+    \defgroup manip_func_flip flip
+    @{
+
+    Flip the input along sepcified dimension
+
+    \ingroup manip_mat
+    \ingroup arrayfire_func
+   */
+    AFAPI array flip(const array &in, const unsigned dim);
+   /**
+      @}
+   */
+
 }
 #endif
 
@@ -292,6 +441,41 @@ extern "C" {
     /**
        @}
     */
+
+    /**
+       \ingroup manip_func_join
+    */
+    AFAPI af_err af_join(af_array *out, const int dim, const af_array first, const af_array second);
+
+    AFAPI af_err af_join_many(af_array *out, const int dim, const unsigned n_arrays, const af_array *inputs);
+
+    /**
+       \ingroup manip_func_tile
+    */
+    AFAPI af_err af_tile(af_array *out, const af_array in,
+                         const unsigned x, const unsigned y, const unsigned z, const unsigned w);
+
+    /**
+       \ingroup manip_func_reorder
+    */
+    AFAPI af_err af_reorder(af_array *out, const af_array in,
+                            const unsigned x, const unsigned y, const unsigned z, const unsigned w);
+
+    /**
+       \ingroup manip_func_shift
+    */
+    AFAPI af_err af_shift(af_array *out, const af_array in, const int x, const int y, const int z, const int w);
+
+    /**
+       \ingroup manip_func_moddims
+    */
+    AFAPI af_err af_moddims(af_array *out, const af_array in, const unsigned ndims, const dim_type * const dims);
+
+    /**
+       \ingroup manip_func_flat
+    */
+    AFAPI af_err af_flat(af_array *out, const af_array in);
+
 
 #ifdef __cplusplus
 }
