+++ src/api/c/assign.cpp
@@ -10,6 +10,7 @@
 #include <af/dim4.hpp>
 #include <af/defines.h>
 #include <af/index.h>
+#include <af/data.h>
 #include <ArrayInfo.hpp>
 #include <err_common.hpp>
 #include <handle.hpp>
