+++ src/api/c/moddims.cpp
@@ -9,7 +9,7 @@
 
 #include <af/dim4.hpp>
 #include <af/defines.h>
-#include <af/index.h>
+#include <af/data.h>
 #include <backend.hpp>
 #include <err_common.hpp>
 #include <handle.hpp>
