+++ src/api/c/transpose.cpp
@@ -9,8 +9,8 @@
 
 #include <af/dim4.hpp>
 #include <af/defines.h>
-#include <af/index.h>
 #include <af/blas.h>
+#include <af/data.h>
 #include <err_common.hpp>
 #include <handle.hpp>
 #include <backend.hpp>
